requirements.txt
# AI Employee — Gold Tier Dependencies
# Install with: pip install -r requirements.txt
# Then run:     python -m playwright install chromium

# ── Core (all watchers) ───────────────────────────────────────────────────────
watchdog>=3.0.0
python-dotenv>=1.0.0

# ── Gmail watcher ─────────────────────────────────────────────────────────────
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0

# ── LinkedIn watcher & orchestrator (HTTP calls) ──────────────────────────────
requests>=2.28.0

# ── Playwright (Twitter, Facebook, Instagram, LinkedIn posters) ───────────────
playwright>=1.40.0

# ── ERPNext watcher (HTTP polling) ───────────────────────────────────────────
httpx>=0.25.0
